requirements.txt
python-telegram-bot==22.4
motor>=3.6.0
pymongo>=4.10.1,<5
python-dotenv==1.0.1
certifi>=2024.7.4
dnspython>=2.6.1

